% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asis.R, R/new.R
\name{is.stack}
\alias{is.stack}
\alias{is.fifo}
\alias{is.lifo}
\alias{is.nino}
\alias{as.list.stack}
\alias{fifo}
\alias{lifo}
\alias{nino}
\title{Stacks - creation and class}
\usage{
is.stack(x)

is.fifo(x)

is.lifo(x)

is.nino(x)

\method{as.list}{stack}(x, ...)

fifo(max_length = Inf, max_size = Inf)

lifo(max_length = Inf, max_size = Inf)

nino(max_length = Inf, max_size = Inf)
}
\arguments{
\item{x}{An object to be tested or coerced.}

\item{...}{Additional arguments.}

\item{max_length}{numeric. The maximum (infinite by default) number of objects 
the stack can contain.}

\item{max_size}{numeric. The maximum (infinite by default) size of
the stack, in octets.}
}
\value{
\code{is.xxx} functions return a logical. 

\code{fifo}, \code{lifo}, and \code{nino} return 
an empty FIFO, LIFO, or NINO stack.
}
\description{
The \code{fifo}, \code{lifo}, and \code{nino} functions 
create 'First In First Out',  'Last In First Out', and 
'Not In or Never Out' stacks, respectively.
}
\seealso{
\code{\link[flifo]{push}}, 
\code{\link[flifo]{pop}}.
}
