% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{height}
\alias{height}
\alias{max_height}
\alias{max_height<-}
\title{Height of a stack}
\usage{
height(.stack)

max_height(.stack)

max_height(x) <- value
}
\arguments{
\item{.stack}{A stack.}

\item{x}{A stack.}

\item{value}{numeric. The new maximum height of the stack.}
}
\value{
\code{height} always returns a nonnegative integer. 
\code{max_height} returns a (possibly infinite) nonnegative numeric.
}
\description{
The function \code{height} returns the height of a stack, 
defined as its number of elements. 
The function \code{max_height} returns the maximum number 
of objects a stack can contains; this number can be changed 
with \code{max_height<-}.
}

