% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.empty.stack.R
\name{is.empty.stack}
\alias{is.empty.stack}
\title{Test emptyness of a stack}
\usage{
\method{is.empty}{stack}(x)
}
\arguments{
\item{x}{A stack.}
}
\value{
A logical, \code{TRUE} if \code{x} is empty.
}
\description{
This method tests if a stack \code{x} is empty.
}
\seealso{
The generic function \code{\link[bazar]{is.empty}} 
in package \pkg{bazar}.
}
