% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_length.R
\name{max_length}
\alias{max_length}
\alias{max_length<-}
\title{Maximum length of a stack}
\usage{
max_length(.stack)

max_length(x) <- value
}
\arguments{
\item{.stack, x}{A stack.}

\item{value}{numeric. The new maximum length of the stack.}
}
\value{
\code{max_length} returns a (possibly infinite) 
nonnegative numeric.
}
\description{
The function \code{max_length} returns the maximum number 
of objects a stack can contains; this number can be changed 
with \code{max_length<-}.
}
