% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stack.R
\name{print.stack}
\alias{print.stack}
\title{Print a stack.}
\usage{
\method{print}{stack}(x, ...)
}
\arguments{
\item{x}{A stack.}

\item{...}{Additional arguments.}
}
\value{
The stack \code{x} is returned invisibly.
}
\description{
The function \code{print.stack} 
prints the class of the stack \code{x} (FIFO, LIFO, or NINO) and displays 
its next reachable object.
}
\seealso{
\code{\link[flifo]{push}}, \code{\link[flifo]{pop}}.
}
