% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\alias{max_size}
\alias{max_size<-}
\title{Size of a stack}
\usage{
size(.stack)

max_size(.stack)

max_size(x) <- value
}
\arguments{
\item{.stack}{A stack.}

\item{x}{A stack.}

\item{value}{numeric. The new maximum size of the stack.}
}
\value{
\code{size} always returns a nonnegative numeric.  
\code{max_size} returns a (possibly infinite) nonnegative numeric.
}
\description{
The function \code{size} returns the size of a stack, in bytes. 
The function \code{max_size} returns the maximum number 
of objects a stack can contains; this number can be changed 
with \code{max_size<-}.
}
